+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] [tooBigData]
+### Fixed
+
+- set asynchrony all create components. Then draw only one component per time. Then get data from rest throw web worker.
+
 ## [1.43.25] [heightTabel]
 ### Fixed